+++ src/senaite/ast/upgrade/v01_00_000.py
@@ -146,15 +146,11 @@ def remove_ast_from_profiles(portal):
     """
     logger.info("Removing AST-like analyses from profiles ...")
     ast_uids = get_ast_services_uids(portal)
-    query = {
-        "portal_type": "AnalysisProfile"
-    }
-    brains = api.search(query, SETUP_CATALOG)
-    for brain in brains:
-        obj = api.get_object(brain)
-        services = obj.getRawService() or []
-        services = filter(lambda s: s not in ast_uids, services)
-        obj.setService(services)
+    profiles = portal.setup.analysisprofiles.objectValues()
+    for obj in profiles:
+        services = obj.getRawServices() or []
+        services = filter(lambda s: s.get("uid") not in ast_uids, services)
+        obj.setServices(services)
     logger.info("Removing AST-like analyses from profiles [DONE]")